+++ equinity-historical-data/src/main/java/uk/gov/justice/laa/crime/equinity/historicaldata/mapper/CrmMapper.java
@@ -1,6 +1,8 @@
 package uk.gov.justice.laa.crime.equinity.historicaldata.mapper;
 
+import org.apache.commons.lang3.StringUtils;
 import org.mapstruct.Mapping;
+import org.mapstruct.Named;
 import uk.gov.justice.laa.crime.equinity.historicaldata.generated.dto.CrmFurtherInformationDTO;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmDecisionReasonDetailsModelInterface;
 import uk.gov.justice.laa.crime.equinity.historicaldata.model.CrmFurtherInfoAttachmentsModel;
@@ -10,6 +12,8 @@ import java.text.ParseException;
 import java.util.Date;
 
 public interface CrmMapper {
+    String DEFAULT_START_DATETIME="1899-12-30T00:00:00";
+    String ISO_DATE_PATTERN="yyyy-MM-dd'T'HH:mm:ss";
     String DECISION_GRANTED = "G";
 
     // Generic converters
@@ -37,6 +41,18 @@ public interface CrmMapper {
     default String convertDecisionReason(CrmDecisionReasonDetailsModelInterface model) {
         return (model.getDecision_original().equals(DECISION_GRANTED)) ?
                 model.getFull_grant_notes() : model.getReason_details();
+    }
+
+    @Named("convertToTimeSpentString")
+    default String convertToTimeSpentString(String endDateTime){
+        if (StringUtils.isEmpty(endDateTime)){
+            return null;
+        }
+        if (endDateTime.length() > 8){
+            return DateUtil.calculateTimeDifference(DEFAULT_START_DATETIME,endDateTime,ISO_DATE_PATTERN);
+        } else {
+            return endDateTime;
+        }
     }
 
     @Mapping(target = "name", source = "name")
